+++ ARCPORTAL-SDK-APP/src/main/java/org/akraino/portal_user_interface/arcportalsdkapp/controller/BlueprintController.java
@@ -16,8 +16,6 @@
 
 package org.akraino.portal_user_interface.arcportalsdkapp.controller;
 
-import org.akraino.portal_user_interface.arcportalsdkapp.client.arc.resources.blueprint.Blueprint;
-import org.akraino.portal_user_interface.arcportalsdkapp.client.arc.resources.blueprint.Blueprints;
 import org.akraino.portal_user_interface.arcportalsdkapp.service.BlueprintService;
 import org.onap.portalsdk.core.controller.RestrictedBaseController;
 import org.onap.portalsdk.core.logging.logic.EELFLoggerDelegate;
@@ -45,34 +43,34 @@ public class BlueprintController extends RestrictedBaseController {
     }
 
     @RequestMapping(value = { "/" }, method = RequestMethod.GET)
-    public ResponseEntity<Blueprints> getBlueprints() {
+    public ResponseEntity getBlueprints() {
         try {
             return new ResponseEntity<>(service.getBlueprints(), HttpStatus.OK);
         } catch (Exception e) {
             LOGGER.error(EELFLoggerDelegate.errorLogger,
-                    "Error occured when trying to retrieve Blueprints. " + UserUtils.getStackTrace(e));
+                    "Error occurred when trying to retrieve blueprints. " + UserUtils.getStackTrace(e));
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
         }
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
     }
 
     @RequestMapping(value = { "/{id}" }, method = RequestMethod.GET)
-    public ResponseEntity<Blueprint> getBlueprint(@PathVariable("id") String uuid) {
+    public ResponseEntity getBlueprint(@PathVariable("id") String uuid) {
         try {
             return new ResponseEntity<>(service.getBlueprint(uuid), HttpStatus.OK);
         } catch (Exception e) {
             LOGGER.error(EELFLoggerDelegate.errorLogger,
-                    "Error when retrieving Blueprint. " + UserUtils.getStackTrace(e));
+                    "Error occurred when trying to retrieve blueprint. " + UserUtils.getStackTrace(e));
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
         }
-        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
     }
 
     @RequestMapping(value = { "/" }, method = RequestMethod.POST)
-    public ResponseEntity<Blueprint> createBlueprint(@RequestBody String blueprint) {
+    public ResponseEntity createBlueprint(@RequestBody String blueprint) {
         try {
             return new ResponseEntity<>(service.saveBlueprint(blueprint), HttpStatus.CREATED);
         } catch (Exception e) {
             LOGGER.error(EELFLoggerDelegate.errorLogger, "Creation of blueprint failed. " + UserUtils.getStackTrace(e));
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
         }
     }
 }